requirements.txt
# Core dependencies for OpenAI Realtime API
websockets==12.0
python-dotenv==1.0.0
aiohttp==3.9.1

# Audio processing
numpy==1.26.2
scipy==1.11.4

# OpenAI SDK (includes async support)
openai==1.40.0

# RAG dependencies
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.22
tiktoken==0.5.2
PyPDF2==3.0.1

# API orchestration
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Flight booking APIs
amadeus==9.0.0

# Utilities
python-json-logger==2.0.7
rich==13.7.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
pylint==3.0.3
mypy==1.8.0